requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alibabacloud-credentials==0.2.0
alibabacloud-ecs20140526==2.1.0
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.1.6
alibabacloud-rds20140815==2.0.5
alibabacloud-tea==0.2.8
alibabacloud-tea-openapi==0.3.1
alibabacloud-tea-util==0.3.5
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.2
Deprecated==1.2.13
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
environs==9.5.0
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
Markdown==3.3.6
marshmallow==3.15.0
multidict==6.0.2
packaging==21.3
prompt-toolkit==3.0.28
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2022.1
redis==4.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
yarl==1.7.2
zipp==3.7.0
